+++ duckscript_sdk/src/sdk/std/env/get_username/mod_test.rs
@@ -9,11 +9,11 @@ fn common_functions() {
 
 #[test]
 fn run_valid() {
-    let name = whoami::hostname();
+    let name = whoami::username();
 
     test::run_script_and_validate(
         vec![create("")],
-        "out = hostname",
+        "out = whoami",
         CommandValidation::Match("out".to_string(), name),
     );
 }