+++ duckscript_sdk/src/sdk/std/env/mod.rs
@@ -1,6 +1,8 @@
 mod cd;
 mod get;
 mod get_home_dir;
+mod get_username;
+mod os_family;
 mod pwd;
 mod set;
 mod unset;
@@ -17,6 +19,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(cd::create(&package))?;
     commands.set(get::create(&package))?;
     commands.set(get_home_dir::create(&package))?;
+    commands.set(get_username::create(PACKAGE))?;
+    commands.set(os_family::create(PACKAGE))?;
     commands.set(pwd::create(&package))?;
     commands.set(set::create(&package))?;
     commands.set(unset::create(&package))?;