+++ duckscript_sdk/src/sdk/std/mod.rs
@@ -13,7 +13,6 @@ mod math;
 mod net;
 mod not;
 pub(crate) mod on_error;
-mod os_family;
 mod process;
 mod read;
 pub(crate) mod release;
@@ -38,7 +37,6 @@ pub(crate) fn load(commands: &mut Commands) -> Result<(), ScriptError> {
     commands.set(is_defined::create(PACKAGE))?;
     commands.set(man::create(PACKAGE))?;
     commands.set(not::create(PACKAGE))?;
-    commands.set(os_family::create(PACKAGE))?;
     commands.set(read::create(PACKAGE))?;
     commands.set(release::create(PACKAGE))?;
     commands.set(set::create(PACKAGE))?;